+++ src/main/java/com/fooddiary/api/dto/response/diary/HomeResponseDTO.java
@@ -3,7 +3,6 @@ package com.fooddiary.api.dto.response.diary;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fooddiary.api.entity.diary.Time;
 import lombok.*;
 
 @Getter
@@ -12,7 +11,7 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class HomeResponseDTO {
-    private int id; // diary id
+    private long id; // diary id
     private byte[] bytes;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private LocalDate time;
