+++ src/main/java/com/fooddiary/api/repository/ImageRepository.java
@@ -15,7 +15,7 @@ public interface ImageRepository extends JpaRepository<Image, Integer> {
   //  Optional<Image> findParentImageByImageIdAndUserId(@Param("imageId") int imageId, @Param("userId") int userId);
 
     @Query("select i from Image i where i.id = :imageId")
-    Optional<Image> findByImageIdAndUserId(@Param("imageId") int imageId);
+    Optional<Image> findByImageIdAndUserId(@Param("imageId") long imageId);
 
    // @Query("select i from Image i left join fetch i.diaryTags T where i.id = :id and i.parentImage.id is null and i.user.id = :userId")
    // Optional<Image> findByIdWithTag(@Param("id") int id, @Param("userId") int userId);
