+++ src/main/java/com/fooddiary/api/repository/diary/DiaryRepository.java
@@ -35,7 +35,7 @@ public interface DiaryRepository extends JpaRepository<Diary, Integer> {
     int getDiaryImagesCount(@Param("id") int id);
 
     @Query("select d from Diary d left join fetch d.images i where d.user.id = :userId and d.id > :id order by d.id asc")
-    List<Diary> findByUserIdAndLimit(@Param("userId") int userId, @Param("id") int id, Pageable pageable);
+    List<Diary> findByUserIdAndLimit(@Param("userId") int userId, @Param("id") long id, Pageable pageable);
 
     Diary findByUserIdAndId(@Param("userId") int userId, @Param("id") int id);
 }
